pyproject.toml
[project]
name = "kwcc-tdz"
version = "0.1.0"
description = "KWCC Tour de Zwift 2026 Results and GC Standings"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "jinja2>=3.1.0",
    "boto3>=1.34.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "boto3-stubs[s3,secretsmanager,cloudfront]>=1.34.0",
    "moto>=5.0.0",
]
cdk = [
    "aws-cdk-lib>=2.130.0",
    "constructs>=10.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
